feature([feature/slide-OurBlog]): solving the carousel problem
