+++ client/src/App.jsx
@@ -1,6 +1,7 @@
 import {
   PostList,
 } from './components';
+import OurBlogSlide from './components/OurBlogSlide/OurBlogSlide'
 import OurBlog from './components/OurBlog/OurBlog'
 
 function App() {
