+++ client/src/components/OurBlog/OurBlog.css
@@ -1,177 +1,17 @@
-.slick-slider{
-	max-width: 70vw;
-  align-items: center;
-	left: 15%;
-	display: block;
-	position: relative;
-}
-.slick-slide {
-  /*text-align: center;*/
-  position: relative;
-}
-
-.slick-prev:before,
-.slick-next:before {
-  color: #777777;
-}
-
-
-.slick-prev{
-	position: absolute;
+@import url('https://fonts.googleapis.com/css2?family=Poppins&display=swap');
+.tudo{
 	display: flex;
-	flex-direction: row;
-	left: 28vw;
-	z-index: 10;
-	
-}
-
-.slick-prev:before{
-	content: url('../Assets/prev.svg');
-	
-}
-
-.slick-next:before{
-	content: url('../Assets/next.svg');
-	
-}
-
-.slick-next{
-	position: absolute;
-	display: flex;
-	flex-direction: row;
-	right: 28vw;
-	z-index: 10;
-}
-
-/*
-.slick-arrow{
-	width: 10px;
-	border-radius: 100px;
-	top: 30vh;
-}
-*/
-
-.slick-prev, .slick-next{
-	
-	height: 0px;
-	top: 101%
-}
-
-.slick-dots{
-	display: flex;
-  align-items: center;
-  justify-content: center;
-  top: auto;
-  padding: 0;
-}
-
-.todo{
-	display: flex;
+	flex-direction: column;
+	background-color: #EBECEE;
 	justify-content: center;
-	max-width: 50%;
-	left: 24vw;
-	margin: 0;
+	height: 100%;
 }
 
-.section{
-	display: flex;
-	flex-direction: row;
-	margin: 0;
-	background-color: aquamarine;
-	/*
-	width: 10vw;
-	height: 30vh;
-	*/
-}
-
-.img{
-	width: 150px;
-  height: 100px;
-  margin: 0;
-  object-fit: cover;
-}
-
-.img-card {
-	display: flex;
-  border-radius: 4px;
-  background: #f5f5f5;
-  color: #666;
-  overflow: hidden;
-  box-shadow: 0 2px 4px 0 rgba(0,0,0,0.2);
-  transition: 0.3s;
-}
-
-.card-body {
-  padding: 15px;
-}
- 
-.card-title {
-  font-size: 20px;
-  font-weight: 600;
-}
- 
-.card-text {
-  margin-top: 10px;
-}
-
-.slick-dots li button:before {
-	content: url('../Assets/RectangleSem.svg');
-}
-
-.slick-dots li.slick-active button:before{
-	content: url('../Assets/RectangleCom.svg');
-}
-
-@media screen and (max-width: 1085px){
-	.slick-next{
-		right: 25vw;
-	}
-
-	.slick-prev{
-		left: 25vw;
-	}
-}
-
-@media screen and (max-width: 768px){
-	.slick-dots{
-		top: auto;
-	}
-
-	.slick-next{
-		right: 20vw;
-	}
-
-	.slick-prev{
-		left: 20vw;
-	}
-}
-
-@media screen and (max-width: 487px){
-	.slick-next{
-		right: 15vw;
-	}
-
-	.slick-prev{
-		left: 15vw;
-	}
-}
-
-@media screen and (max-width: 349px){
-	.slick-next{
-		right: 10vw;
-	}
-
-	.slick-prev{
-		left: 10vw;
-	}
-}
-
-@media screen and (max-width: 280px){
-	.slick-next{
-		right: 8vw;
-	}
-
-	.slick-prev{
-		left: 8vw;
-	}
+.tudo h1{
+	font-family: 'Poppins', sans-serif;
+	align-items: center;
+	font-weight: 700;
+	color: #032C66;
+	text-align: center;
+	margin-top: auto;
 }
