--- client/src/components/OurBlog/data.jsx
@@ -1,29 +1,0 @@
-import Img01 from '../Assets/img01.jpg'
-import Img02 from '../Assets/img02.jpg'
-
-export default [
-	{
-		img: Img01,
-		title: 'Título do blogpost',
-		data: 'Data da publicação',
-		text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam...'
-	},
-	{
-		img: Img02,
-		title: 'Título do blogpost',
-		data: 'Data da publicação',
-		text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam...'
-	},
-	{
-		img: Img01,
-		title: 'Título do blogpost',
-		data: 'Data da publicação',
-		text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam...'
-	},
-	{
-		img: Img02,
-		title: 'Título do blogpost',
-		data: 'Data da publicação',
-		text: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam...'
-	},
-]
