+++ client/src/components/OurBlogSlide/OurBlogSlide.css
@@ -1,0 +1,130 @@
+
+.slick-slider{
+	max-width: 70vw;
+  align-items: center;
+	left: 15%;
+	display: block;
+	position: relative;
+}
+.slick-slide {
+  /*text-align: center;*/
+  position: relative;
+}
+
+.slick-prev:before,
+.slick-next:before {
+  color: #777777;
+}
+
+
+.slick-prev{
+	position: absolute;
+	display: flex;
+	flex-direction: row;
+	left: 28vw;
+	z-index: 10;
+	
+}
+
+.slick-prev:before{
+	content: url('../Assets/prev.svg');
+	
+}
+
+.slick-next:before{
+	content: url('../Assets/next.svg');
+	
+}
+
+.slick-next{
+	position: absolute;
+	display: flex;
+	flex-direction: row;
+	right: 28vw;
+	z-index: 10;
+}
+
+/*
+.slick-arrow{
+	width: 10px;
+	border-radius: 100px;
+	top: 30vh;
+}
+*/
+
+.slick-prev, .slick-next{
+	
+	height: 0px;
+	top: 101%
+}
+
+.slick-dots{
+	display: flex;
+  align-items: center;
+  justify-content: center;
+  top: auto;
+  padding: 0;
+}
+
+
+.slick-dots li button:before {
+	content: url('../Assets/RectangleSem.svg');
+}
+
+.slick-dots li.slick-active button:before{
+	content: url('../Assets/RectangleCom.svg');
+}
+
+@media screen and (max-width: 1085px){
+	.slick-next{
+		right: 25vw;
+	}
+
+	.slick-prev{
+		left: 25vw;
+	}
+}
+
+@media screen and (max-width: 768px){
+	.slick-dots{
+		top: auto;
+	}
+
+	.slick-next{
+		right: 20vw;
+	}
+
+	.slick-prev{
+		left: 20vw;
+	}
+}
+
+@media screen and (max-width: 487px){
+	.slick-next{
+		right: 15vw;
+	}
+
+	.slick-prev{
+		left: 15vw;
+	}
+}
+
+@media screen and (max-width: 349px){
+	.slick-next{
+		right: 10vw;
+	}
+
+	.slick-prev{
+		left: 10vw;
+	}
+}
+
+@media screen and (max-width: 280px){
+	.slick-next{
+		right: 8vw;
+	}
+
+	.slick-prev{
+		left: 8vw;
+	}
+}
